pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arxiv-newsletter"
version = "0.1.0"
description = "A personalized newsletter generator for arXiv papers"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["arxiv", "newsletter", "research", "papers"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "arxiv>=2.0.0",
    "requests>=2.28.0",
    "feedparser>=6.0.0",
    "reportlab>=4.0.0",
    "PyYAML>=6.0",
    "numpy>=1.21.0",
    "scikit-learn>=1.0.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
]

[project.scripts]
arxiv-newsletter = "arxiv_newsletter.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["arxiv_newsletter*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
